% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_add_fc_values_per_split_date.R
\name{add_fc_values_per_split_date}
\alias{add_fc_values_per_split_date}
\title{Add computed hierarchical forecast values, per split date}
\usage{
add_fc_values_per_split_date(
  main_forecasting_table,
  grouped_hierarchical_fc_values
)
}
\arguments{
\item{main_forecasting_table}{A tibble object that contains run forecasts for
one split date and all groups. It is assumed that this is the output of the
\code{add_fc_models_to_main_forecasting_table} or
\code{update_main_forecasting_table} function.}

\item{grouped_hierarchical_fc_values}{A tibble that containts the best
fc_model for each group. It is assumed that this tibble is created using
the \code{get_forecast_accuracy_overview} and
\code{get_best_forecast_methods} functions (see example).}
}
\value{
A tibble with the consistent hierarchical forecast value for all
  groups in all forecast dates, at the specified forecast split date
}
\description{
\code{add_fc_values_per_split_date} is a function that adds previously
computed hierarchical forecasts into the main forecasting table, for a single
split date.
}
\examples{
main_forecasting_table <- dummy_hierarchical_gasprice \%>\%
  dplyr::filter(oil_company == "CompanyA") \%>\%
  tstools::initialize_ts_forecast_data(
    date_col = "year_month",
    col_of_interest = "gasprice",
    group_cols = "currency",
    hierarchical_cols = c("location")
  ) \%>\%
  dplyr::filter(period >= as.Date("2004-06-30")) \%>\%
  create_main_forecasting_table() \%>\%
  dplyr::filter(ts_split_date == 200605) \%>\%
  add_fc_models_to_main_forecasting_table(
    fc_methods = c("basic", "linear")
  )
best_model_per_group <- main_forecasting_table \%>\%
  get_forecast_accuracy_overview() \%>\%
  get_best_forecast_methods()
grouped_hierarchical_fc_values <- get_consistent_fc_values(
  main_forecasting_table = main_forecasting_table,
  best_model_per_group = best_model_per_group
)
add_fc_values_per_split_date(
  main_forecasting_table = main_forecasting_table,
  grouped_hierarchical_fc_values = grouped_hierarchical_fc_values
)
}
