% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTRA_FORECAST_add_kalman_forecast_model.R
\name{add_kalman_forecast_model}
\alias{add_kalman_forecast_model}
\title{Add kalman forecast model}
\usage{
add_kalman_forecast_model(
  fc_models,
  ts_object_train,
  fc_name,
  periods_ahead = 1,
  periods_history = Inf,
  verbose = FALSE,
  log_message = ""
)
}
\arguments{
\item{fc_models}{A named list of forecast models, with for each forecast
model a list with the model itself and a table with forecast values.}

\item{ts_object_train}{A time series object, which contains only the training
data.}

\item{fc_name}{A character string specifying the name to be used for the new
model that is added to the list of existing forecast models.}

\item{periods_ahead}{A positive integer value indicating the number of
periods to forecast ahead.}

\item{periods_history}{A positive integer value indicating the number of
historic datapoints to use for training, which is only relevant for
specific forecast methods such as drift and mean.}

\item{verbose}{Boolean, which is set to TRUE if status updates are valued, or
set to FALSE if they are not.}
}
\value{
A named list of forecast models, with for each forecast model a list
  with the model itself and a table with forecast values.
}
\description{
\code{add_kalman_forecast_model} is a function to add a single kalman
forecast model to a (named) list of forecast models.
}
\examples{
ts_object_train <- tstools::initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = c("state", "oil_company")
   ) \%>\%
   dplyr::filter(grouping == "state = New York   &   oil_company = CompanyA") \%>\%
   tstools::transform_data_to_ts_object()
add_kalman_forecast_model(
   fc_models = list(),
   ts_object_train = ts_object_train,
   fc_name = "fc_kalman_poly",
   periods_ahead = 12,
   verbose = T
)    
}
