% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_FORECAST_add_mean_ensemble_fc_methods.R
\name{add_mean_ensemble_fc_methods}
\alias{add_mean_ensemble_fc_methods}
\title{Adds an mean ensemble of provided forecast methods.}
\usage{
add_mean_ensemble_fc_methods(
  main_forecasting_table,
  fc_models,
  fc_model_name = "fc_mean_ensemble"
)
}
\arguments{
\item{main_forecasting_table}{A tibble containing a single row and several
columns of data required for time series forecasting, which has been
created using the \code{create_main_forecasting_table} function and which
has been extended with the fc_models and fc_errors columns using the
\code{add_fc_models_to_main_forecasting_table} function.}

\item{fc_models}{A character vector specifying which forecast models to use
in the ensemble.}

\item{fc_model_name}{A character string specifying the name of the mean
ensamble method that is added to the main_forecasting_table.}
}
\value{
The main_forecasting_table with a mean ensemble method added.
}
\description{
\code{add_mean_ensemble_fc_methods} is a function to add a new forecast
method after other forecast methods have been added to the
main_forecast_table. It takes the average value of the given forecast methods
as the forecast value.
}
\examples{
tstools::initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = c("state", "oil_company")
   ) \%>\%
   create_main_forecasting_table() \%>\%
   head(1) \%>\%
   add_fc_models_to_main_forecasting_table(
      periods_ahead = 12,
      fc_methods = c("linear")
   ) \%>\%
   add_mean_ensemble_fc_methods(
      fc_models = c("fc_linear_trend", "fc_linear_trend_seasonal")
   )
}
