% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INTRA_FORECAST_add_prophet_forecast_model.R
\name{add_prophet_forecast_model}
\alias{add_prophet_forecast_model}
\title{Add prophet forecast model}
\usage{
add_prophet_forecast_model(
  fc_models,
  ts_object_train,
  ts_object_valid = NULL,
  fc_name,
  model_type = c("univariate", "multivariate"),
  changepoint_prior_scale,
  periods_ahead = 1,
  periods_history = Inf,
  keep_fc_model_objects = FALSE,
  verbose = FALSE,
  log_message = ""
)
}
\arguments{
\item{fc_models}{A named list of forecast models, with for each forecast
model a list with the model itself and a table with forecast values.}

\item{ts_object_train}{A time series object, which contains only the training
data.}

\item{ts_object_valid}{A time series object, which contains the validation
data. This is used for multivariate frameworks, thus it should have the
forecasted/actual values of the external regressors as well.}

\item{fc_name}{A character string specifying the name to be used for the new
model that is added to the list of existing forecast models.}

\item{model_type}{A character string indicating whether a univariate model
(without external regressors) or a multivariate model (with external
regressors) should be estimated.}

\item{changepoint_prior_scale}{A positive numeric value modulating the
flexibility of the automatic changepoint selection, where large values will
allow many changepoints, while small values will allow few changepoints.}

\item{periods_ahead}{A positive integer value indicating the number of
periods to forecast ahead.}

\item{periods_history}{A positive integer value indicating the number of
historic datapoints to use for training, which is only relevant for
specific forecast methods such as drift and mean.}

\item{verbose}{Boolean, which is set to TRUE if status updates are valued, or
set to FALSE if they are not.}
}
\value{
A named list of forecast models, with for each forecast model a list
  with the model itself and a table with forecast values.
}
\description{
\code{add_prophet_forecast_model} is a function to add a single prophet
forecast model to a (named) list of forecast models. The forecast model is
created based on a parameter value to determine the flexibility of automatic
changepoint selection, which is then used to forecast a specific number of
periods ahead.
}
\examples{
ts_object_train <- tstools::initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = c("state", "oil_company"),
      xreg_cols = c("spotprice", "gemprice")
   ) \%>\%
   dplyr::filter(grouping == "state = New York   &   oil_company = CompanyA") \%>\%
   tstools::transform_data_to_ts_object()
add_prophet_forecast_model(
   fc_models = list(),
   ts_object_train = ts_object_train,
   fc_name = "fc_prophet_050cps",
   changepoint_prior_scale = 0.050,
   periods_ahead = 12,
   verbose = T
)    
}
