% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsforecast_1.3.1.R
\docType{data}
\name{airport}
\alias{airport}
\title{Airport Travellers Time Series.}
\format{
A data frame with 84 rows and 2 variables:
\describe{
  \item{Date}{Months in which the traveller numbers are recorded.}
  \item{Travellers}{Total travellers in the airport of that month.}
  \item{AvgTemp}{Monthly average temperature (in °C) of the city where the airport is located.}
  \item{AvgRain}{Monthly average rain fall (in mm) in the city where the airport is located.}
}
}
\usage{
airport
}
\description{
A dataset containing the total number of monthly travellers recorded
in an unnamed airport between January 2013 and December 2019.
}
\examples{
data(airport)
head(airport)
}
\keyword{datasets}
