% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_change_ts_object_seasonality.R
\name{change_ts_object_seasonality}
\alias{change_ts_object_seasonality}
\title{Add or modify the seasonality attribute of a time series object}
\usage{
change_ts_object_seasonality(ts_object, seasonality)
}
\arguments{
\item{ts_object}{A time series object containing the values for the
column_of_interest, which has been created using the
\code{transform_data_to_ts_object} from tstools, or alternatively an empty list.}

\item{seasonality}{A numerical vector indicating the seasonality to employ
for the time series object, for example c(12, 3).}
}
\value{
A time series object with the specified seasonality.
}
\description{
\code{change_ts_object_seasonality} is a function to modify or add a
seasonality attribute to an existing time series object.
}
\examples{
ts_object <- tstools::initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = c("state", "oil_company"),
      xreg_cols = c("spotprice", "gemprice")
   )\%>\%
   dplyr::filter(grouping == "state = New York   &   oil_company = CompanyB") \%>\%
   tstools::transform_data_to_ts_object(seasonal_periods = 1)
change_ts_object_seasonality(ts_object, seasonality = c(12, 3, 4))
}
