% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_cleanup_multivariate_ts_object.R
\name{cleanup_multivariate_ts_object}
\alias{cleanup_multivariate_ts_object}
\title{Cleans a multivariate time series object}
\usage{
cleanup_multivariate_ts_object(ts_object)
}
\arguments{
\item{ts_object}{A multivariate time series object}
}
\value{
A cleaned multivariate time series object.
}
\description{
\code{cleanup_multivariate_ts_object} is a function to clean a multivariate
time series object by removing all unnecessary columns (e.g.
original_col_of_interest). This is needed to make sure that no information
besides the external regressors is used when training the forecast models.
}
\examples{
ts_object <- tstools::initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = c("state", "oil_company"),
      xreg_cols = c("spotprice", "gemprice")
   ) \%>\%
   dplyr::filter(grouping == "state = New York   &   oil_company = CompanyA") \%>\%
   dplyr::mutate(original_col_of_interest = col_of_interest) \%>\% 
   tstools::transform_data_to_ts_object(seasonal_periods = c(12,3))
cleanup_multivariate_ts_object(ts_object = ts_object)
}
