% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VISUALIZE_compare_forecast_per_group.R
\name{compare_forecast_per_group}
\alias{compare_forecast_per_group}
\title{Create plot to compare forecast per group}
\usage{
compare_forecast_per_group(
  main_forecasting_table,
  fc_model = "",
  demo_mode = FALSE
)
}
\arguments{
\item{main_forecasting_table}{A tibble containing a single row per group and
several columns of data required for time series forecasting, which has
been created using the \code{create_main_forecasting_table} function and
which has been extended with the fc_models and fc_errors columns using the
\code{add_fc_models_to_main_forecasting_table} function.}

\item{fc_model}{A character string specifying which forecast model to compare
accross the available groups.}

\item{demo_mode}{Boolean, which is to be set to TRUE if any potentially
sensitive figures should be hidden from the audience for demo purposes, or
set to FALSE if all figures can safely be displayed.}
}
\value{
A plotly object displaying a comparison of forecasts per group.
}
\description{
\code{compare_forecast_per_group} is a function to create a plot which
compares the forecasted values of a single forecast model for every available
group. For demo purposes, sensitive figures can be hidden from the audience.
}
\examples{
tstools::initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = c("state", "oil_company")
   ) \%>\%
   create_main_forecasting_table() \%>\%
   dplyr::filter(ts_split_date == 200503) \%>\%
   add_fc_models_to_main_forecasting_table(
      periods_ahead = 12,
      fc_methods = c("linear")
   ) \%>\%
   compare_forecast_per_group(
      fc_model = "fc_linear_trend_seasonal"
   )
}
