% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_decide_on_parallel_run.R
\name{decide_on_parallel_run}
\alias{decide_on_parallel_run}
\title{Determine whether a parallel run will be faster}
\usage{
decide_on_parallel_run(fc_methods = supported_fc_methods_uni_var(), nrows)
}
\arguments{
\item{fc_methods}{A character vector specifying the forecast methods to run. 
For more info \code{`?supported_fc_methods`}.}

\item{nrows}{A positive integer value indicating the number of rows in the
main_forecasting_table that needs to be filled with forecasts.}
}
\value{
A boolean indicating whether a parallel run is expecting to be faster
  (TRUE) or not (FALSE)
}
\description{
\code{decide_on_parallel_run} is a function to determine whether running the
forecasting in parallel is going to be faster that not running it parallel.
Because of the initialization time of parallel forecast runs, it could be
faster for limited forecasting efforts to run without parallelization.
}
\examples{
decide_on_parallel_run(fc_methods = c("basic"), nrows = 100)
decide_on_parallel_run(fc_methods = c("basic", "prophet"), nrows = 100)
}
