% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_filter_stationary_columns.R
\name{filter_stationary_columns}
\alias{filter_stationary_columns}
\title{Filter stationary columns from the data}
\usage{
filter_stationary_columns(data)
}
\arguments{
\item{data}{A data.table, data.frame or tibble object containing the data to
be used for modelling, from which the stationary columns need to be
filtered.}
}
\value{
A tibble with only non-stationary columns.
}
\description{
\code{filter_stationary_columns} is a function to filter columns which have
only a single unique value from the data. This is done mostly before
modelling, because these stationary columns contain no information that can
be used to create a model.
}
\examples{
dummy_gasprice \%>\% 
   filter_stationary_columns() \%>\% 
   colnames()
dummy_gasprice \%>\% 
   dplyr::filter(state == "New York") \%>\% 
   filter_stationary_columns() \%>\% 
   colnames()
}
