% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/META_generate_training_data_from_M4.R
\name{generate_training_data_from_M4}
\alias{generate_training_data_from_M4}
\title{Generate training data}
\usage{
generate_training_data_from_M4(
  n = 1,
  dir_fforma = file.path(getwd(), "data/fforma_training_data"),
  verbose = TRUE
)
}
\arguments{
\item{n}{A positive integer value indicating the number of sample time series
to draw from the total set of M4 times series.}

\item{dir_fforma}{A character string specifying the path to the directory
where the training data is stored to do FFORMA training.}

\item{verbose}{Boolean, which is set to TRUE if status updates are valued, or
set to FALSE if they are not.}
}
\value{
Nothing.
}
\description{
\code{generate_training_data_from_M4} is a function to generate training data
from the M4 time series forecasting competition data. The M4 data comes from
the M4comp2018 package, which contains 100.000 time series from different
domains. The available M4 time series are filtered to only the monthly time
series, corresponding to Finance data, which are between 5 and 15 years long.
The training data is used for training a Feature-based Forecast Model
Averaging (FFORMA) model.
}
\examples{
generate_training_data_from_M4()
}
