% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VISUALIZE_plot_xreg.R
\name{plot_xreg}
\alias{plot_xreg}
\title{Create plot to compare select external regressors to variable of interest}
\usage{
plot_xreg(main_forecasting_table, xreg = "")
}
\arguments{
\item{main_forecasting_table}{A tibble containing a single row per group and
several columns of data required for time series forecasting, which has
been created using the \code{create_main_forecasting_table} function and
which has been extended with the fc_models and fc_errors columns using the
\code{add_fc_models_to_main_forecasting_table} function. Note that this
table should have the output of a multivariate analysis, otherwise the
function does not run.}

\item{xreg}{A vector that contains strings with the names of the external
regressors to be plotted.}
}
\value{
A plotly object displaying multiple plots.
}
\description{
\code{plot_xreg} is a function to create a plot which compares selected
external regressors with the variable of interest
}
\examples{
tstools::initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = c("state", "oil_company"),
      xreg_cols = c("spotprice", "gemprice")
   ) \%>\%
   create_main_forecasting_table() \%>\%
   dplyr::filter(ts_split_date == 200503) \%>\%
   add_fc_models_to_main_forecasting_table(
      periods_ahead = 12,
      fc_methods = c("linear")
   ) \%>\%
   dplyr::filter(grouping == "state = New York   &   oil_company = CompanyA") \%>\%
   plot_xreg(
      xreg = "spotprice"
   )
}
