% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRAPPER_run_example.R
\name{run_example}
\alias{run_example}
\title{Run example of package}
\usage{
run_example(
  dataset = c("AirPassengers", "nottem", "UKDriverDeaths", "hierarchical"),
  test_mode = F
)
}
\arguments{
\item{dataset}{A character string indicating which of the available public
(time series) datasets to show in the example.}

\item{test_mode}{Boolean, which is only set to TRUE when testing this
function within the testthat framework for building the tsforecast
package.}
}
\value{
An interactive Shiny dashboard.
}
\description{
\code{run_example} is a function to run an example of the application of the
tsforecast package on a public (time series) dataset. Calling the function
will start an interactive Shiny dashboard that enables exploration of the
individual forecasts versus actuals per split date, as well as an analysis of
the overall forecast accuracy of the different forecast models.
}
\examples{
run_example(dataset = "AirPassengers")
run_example(dataset = "nottem")
run_example(dataset = "UKDriverDeaths")
run_example(dataset = "hierarchical")
}
