% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_show_documentation.R
\name{show_documentation}
\alias{show_documentation}
\title{Show package documentation}
\usage{
show_documentation(topic = c("README", "forecast_models"), test_mode = F)
}
\arguments{
\item{topic}{A character string indicating which of the available
documentation files to display.}

\item{test_mode}{Boolean, which is only set to TRUE when testing this
function within the testthat framework for building the tsforecast
package.}
}
\value{
An html document with documentation.
}
\description{
\code{show_documentation} is a function to open the documentation (available
as html file) in a browser.
}
\examples{
show_documentation()
}
