% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/META_store_main_forecasting_table_for_fforma.R
\name{store_main_forecasting_table_for_fforma}
\alias{store_main_forecasting_table_for_fforma}
\title{Sample main forecasting table for package validation}
\usage{
store_main_forecasting_table_for_fforma(
  main_forecasting_table,
  name = "",
  dir_fforma = "//Ad.ing.net/wps/NL/P/GD/200353/Models/FFORMA",
  overwrite = FALSE
)
}
\arguments{
\item{main_forecasting_table}{A tibble containing several columns of data
required for time series forecasting, which has been created using the
\code{create_main_forecasting_table} function and which has been extended
with the fc_models and fc_errors columns using the
\code{add_fc_models_to_main_forecasting_table} function.}

\item{name}{A character string specifying the name of the training data that
is generated from the the main_forecasting_table, to be used for training
FFORMA.}

\item{dir_fforma}{A character string specifying the path to the directory
where the training data is stored to do FFORMA training.}

\item{overwrite}{Boolean, which is set to TRUE if the file with the specified
name should be overwritten if it already exists, or set to FALSE if it is
not.}
}
\value{
Nothing.
}
\description{
\code{store_main_forecasting_table_for_fforma} is a function to sample a
subset of the available rows from the main_forecasting_table, to be used in
the validation of the tsforecast package.
}
\details{
The function will, for each available grouping, get the first number of rows
specified and the last number of rows specified (to test the edge cases) as
well as a random sample of the rows in between of the specified size.
}
\examples{
main_forecasting_table <- dummy_gasprice \%>\%
     tstools::initialize_ts_forecast_data(
     date_col = "year_month",
     col_of_interest = "gasprice",
     group_cols = c("state", "oil_company")
   ) \%>\%
   create_main_forecasting_table(
      seasonal_periods = NULL,
      min_train_periods = 188
   ) \%>\%
   add_fc_models_to_main_forecasting_table(
      fc_methods = c("basic", "linear")
   )
store_main_forecasting_table_for_fforma(
   main_forecasting_table = main_forecasting_table,
   name = "dummy_gasprice",
   dir_fforma = tempdir()
)
}
