% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_sum_ts_objects.R
\name{sum_ts_objects}
\alias{sum_ts_objects}
\title{Sum two ts objects to get a single ts object}
\usage{
sum_ts_objects(ts_objects, new_grouping = "")
}
\arguments{
\item{ts_objects}{A list of time series object to be summed.}

\item{new_grouping}{A character vector containing the grouping of the newly
created time series object.}
}
\value{
A summed time series object.
}
\description{
\code{sum_ts_objects} is a function to sum two separate time series object to
get a single ts object.
}
\examples{
ts_object <- tstools::initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = c("state", "oil_company"),
      xreg_cols = c("spotprice", "gemprice")
   ) \%>\%
   dplyr::filter(grouping == "state = New York   &   oil_company = CompanyA") \%>\%
   tstools::transform_data_to_ts_object(seasonal_periods = 3)
sum_ts_objects(
   ts_objects = list(
      ts_object,
      ts_object,
      ts_object
   ),
   new_grouping = "state = Trice New York   &   oil_company = Trice CompanyA"
)
}
