% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/META_train_fforma_model.R
\name{train_fforma_model}
\alias{train_fforma_model}
\title{Train the FFORMA model}
\usage{
train_fforma_model(
  dir_fforma = file.path(getwd(), "data/fforma_training_data"),
  verbose = TRUE
)
}
\arguments{
\item{dir_fforma}{A character string specifying the path to the directory
where the training data os stored to do FFORMA training.}

\item{verbose}{Boolean, which is set to TRUE if status updates are valued, or
set to FALSE if they are not.}
}
\value{
A FFORMA meta learning model, trained according to the approach
  descibed in \href{https://robjhyndman.com/papers/fforma.pdf}{this paper}
  and
  \href{https://github.com/robjhyndman/M4metalearning/blob/master/docs/metalearning_example.md}{this
   example} on github.
}
\description{
\code{train_fforma_model} is a function to train a Feature-based Forecast
Model Averaging (FFORMA) model, which weights the forecasts from the
different available forecast models within the tsforecast package to
create a new ensamble forecast. The FFORMA model uses the characteristics
from the time series objects to predict which forecast model gives the best
forecast performance. After training on all available training sets within
the specified directory, the resulting FFORMA model is stored within the
package to be reused at runtime during time series forecasting.
}
\examples{
train_fforma_model()
}
