% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_ts_object_to_periods.R
\name{ts_object_to_periods}
\alias{ts_object_to_periods}
\title{Gets available periods from time series object}
\usage{
ts_object_to_periods(ts_object)
}
\arguments{
\item{ts_object}{A time series object.}
}
\value{
A vector of periods in yyyymm numeric format.
}
\description{
\code{ts_object_to_periods} is a function to transform a given time series 
object into a vector of periods in yyyymm numeric format.
}
\examples{
ts_object <- tstools::initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice", 
      group_cols = c("state", "oil_company"), 
      xreg_cols = c("spotprice", "gemprice")
   ) \%>\% 
   dplyr::filter(grouping == "state = New York   &   oil_company = CompanyA") \%>\% 
   tstools::transform_data_to_ts_object(seasonal_periods = 3)
ts_object_to_periods(ts_object)
}
