% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_ts_object_to_tibble.R
\name{ts_object_to_tibble}
\alias{ts_object_to_tibble}
\title{Converts time series object into tibble}
\usage{
ts_object_to_tibble(ts_object)
}
\arguments{
\item{ts_object}{A time series object.}
}
\value{
A tibble object with a period column in yyyymm numeric format and or
  more other columns corresponding to the data in the time series object.
}
\description{
\code{ts_object_to_tibble} is a function to convert a time series object into
a tibble with a period column in yyyymm numeric format and a column of
interest for the data in the time series object. If applicable, also the
external regressors in the time series object are converted to individual
columns.
}
\examples{
ts_object <- tstools::initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = c("state", "oil_company"),
      xreg_cols = c("spotprice", "gemprice")
   ) \%>\%
   dplyr::filter(grouping == "state = New York   &   oil_company = CompanyA") \%>\%
   tstools::transform_data_to_ts_object(seasonal_periods = 3)
ts_object_to_tibble(ts_object = ts_object)
}
