% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HELPER_union_ts_objects.R
\name{union_ts_objects}
\alias{union_ts_objects}
\title{Union two ts objects into a single ts object}
\usage{
union_ts_objects(ts_object_1, ts_object_2)
}
\arguments{
\item{ts_object_1}{The first time series object to be unioned with the
second.}

\item{ts_object_2}{The second time series object to be unioned with the
first.}
}
\value{
A unioned time series object.
}
\description{
\code{union_ts_objects} is a function to union two separate time series
object into a single ts object, which is comparable to a row bind.
}
\examples{
ts_object <- tstools::initialize_ts_forecast_data(
      data = dummy_gasprice,
      date_col = "year_month",
      col_of_interest = "gasprice",
      group_cols = c("state", "oil_company"),
      xreg_cols = c("spotprice", "gemprice")
   ) \%>\%
   dplyr::filter(grouping == "state = New York   &   oil_company = CompanyA") \%>\%
   tstools::transform_data_to_ts_object(seasonal_periods = 3)
ts_part_one <- ts_object \%>\%
   trim_ts_object(
      max_length = 179,
      from_left = T
   )
ts_part_two <- ts_object \%>\%
   trim_ts_object(
      max_length = 12,
      from_left = F
   )
ts_unioned <- union_ts_objects(
   ts_object_1 = ts_part_one,
   ts_object_2 = ts_part_two
)
}
