% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRAPPER_update_main_forecasting_table.R
\name{update_main_forecasting_table}
\alias{update_main_forecasting_table}
\title{Update main forecasting table.}
\usage{
update_main_forecasting_table(
  file_path = "",
  data,
  seasonal_periods = c(12, 3),
  min_train_periods = 25,
  max_train_periods = Inf,
  periods_ahead = 12,
  fc_methods = supported_fc_methods_uni_var(),
  overwrite_fc = FALSE,
  allow_negative_fc = FALSE,
  add_hierarchical_fc = FALSE,
  verbose = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{file_path}{A character string specifying a valid file path.}

\item{data}{A tibble containing the data to be used for time series
forecasting, which has been created using the
\code{initialize_ts_forecast_data} function from tstools.}

\item{seasonal_periods}{A vector of postive integer values indicating the
number of data points that together compose a season (e.g. c(12,3) for
quarterly and yearly seasonality when using monthly data).}

\item{min_train_periods}{A positive integer value indicating the minimum
number of periods of data required for the training time series objects.}

\item{max_train_periods}{A positive integer value indicating the maximum
number of periods of data to be used for the training time series objects.}

\item{periods_ahead}{A positive integer value indicating the number of
periods to forecast ahead.}

\item{fc_methods}{A character vector specifying the forecast methods to add.
For more info \code{`?supported_fc_methods`}.}

\item{overwrite_fc}{Boolean, which is set to TRUE if all existing forecasts
should be overwritten, or set to FALSE if only unavailable forecasts should
be added.}

\item{allow_negative_fc}{Boolean, which is to be set to TRUE if negative
forecast values are allowed, or set to FALSE if negative forecast values
should be overwritten by a value of zero.}

\item{verbose}{Boolean, which is set to TRUE if status updates are valued, or
set to FALSE if they are not.}
}
\value{
An updated tibble containing several columns of data required for
  time series forecasting, as well as a fc_models column with named lists
  with forecast models and forecast data, extended with a fc_errors column
  with the forecast values, actuals and resulting forecast errors.
}
\description{
\code{update_main_forecasting_table} is a wrapper function around the
following functions to create, fill and/or update the main forecasting table
and then store it on disk to prevent recalculation:
\itemize{\item{\code{create_main_forecasting_table}}
\item{\code{add_fc_models_to_main_forecasting_table}}
\item{\code{fast_write_RData_to_network}} from tstools}
}
\examples{
data <- tstools::initialize_ts_forecast_data(
   data = dummy_gasprice,
   date_col = "year_month",
   col_of_interest = "gasprice",
   group_cols = c("state", "oil_company"),
   xreg_cols = c("spotprice", "gemprice")
)
update_main_forecasting_table(
   file_path = file.path(tempdir(), "main_forecasting_table.RData"),
   data = data,
   verbose = T
)
}
