% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsforecast_1.3.1.R
\name{tslag}
\alias{tslag}
\title{Lag a Time Series}
\usage{
tslag(x, lag = 1L, return.matrix = FALSE)
}
\arguments{
\item{x}{a univariate time series object.}

\item{lag}{number of lags (in units of observations). Default is \code{1}. If multiple values are given as a vector, multiple lagged version of the series will be returned as a matrix.}

\item{return.matrix}{optional. A boolean value indicating whether the lagged version of `\code{x}` should be returned as a matrix. Only meaningful if \code{lag} has only one value.}
}
\value{
The same time series object as \code{x} after being back- or foreshifted for the number of periods specified in \code{lag}.
}
\description{
The function `\code{tslag}` back- or foreshifts a time series and generates the lagged version of it.
}
\details{
Note the sign of \code{lag}: a series lagged by a positive \code{lag} starts earlier.
}
\examples{
tslag(airport$Travellers, lag = 12)
}
\seealso{
\link{lag}
}
\author{
Ka Yui Karl Wu
}
